third_party/blink/web_tests/virtual/layout_ng_table/external/wpt/css/css-tables/html5-table-formatting-2-expected.txt

This is a testharness.js-based test.
PASS Explicitely defined columns are not merged
PASS Explicitely defined rows are not merged
PASS Border-spacing is added between any two unmerged columns (1)
PASS Border-spacing is added between any two unmerged rows (1)
PASS Border-spacing is added between any two unmerged columns (2)
PASS Border-spacing is added between any two unmerged rows (2)
PASS Border-spacing is added between any two unmerged columns (3)
PASS Border-spacing is added between any two unmerged rows (3)
PASS Border-spacing is added between any two unmerged columns (4)
PASS Border-spacing is added between any two unmerged rows (4)
PASS Border-spacing is added between any two unmerged columns (5)
PASS Border-spacing is added between any two unmerged rows (5)
Harness: the test ran to completion.
